-Fixed a bunch of warnings(reported by Visual C++). Argh, I'll do the rest another day.
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1305 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/m68kd.c
@@ -1311,7 +1311,7 @@ u32 M68KDisasm(u32 addr, char *outstring)
 
    for (i = 0; instruction[i].name != NULL; i++)
    {
-      u16 op = c68k_word_read(addr);
+      u16 op = (u16)c68k_word_read(addr);
 
       if ((op & instruction[i].mask) == instruction[i].inst)
       {
